refactor: split validate into modules, add issue codes, serde JSON
Split monolithic validate.rs (1700 lines) into modular architecture:
- validate/types.rs — public types, IssueCode enum, helpers
- validate/structure.rs — Level 1 raw byte walking
- validate/metadata.rs — Level 2 CBOR + descriptor checks
- validate/integrity.rs — Level 3 hash + decompression
- validate/mod.rs — public API + tests

Add stable machine-readable IssueCode enum (~40 codes) with
serde Serialize. All ValidationIssue now carry a code field
alongside the human-readable description.

Replace hand-built JSON in CLI with serde_json serialization.
Batch JSON output is now a valid JSON array.

Update TODO.md PR 2 scope with reviewed design decisions.
