+++ crates/tensogram-cli/Cargo.toml
@@ -18,6 +18,7 @@ tensogram-grib = { path = "../tensogram-grib", optional = true }
 tensogram-netcdf = { path = "../tensogram-netcdf", optional = true }
 ciborium.workspace = true
 clap.workspace = true
+serde.workspace = true
 serde_json.workspace = true
 tracing.workspace = true
 tracing-subscriber = { version = "0.3", features = ["env-filter"] }
